requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API clients
openai>=1.0.0
google-cloud-language>=2.11.0

# Scientific computing
scipy>=1.11.0

# Development and analysis
jupyter>=1.0.0
tqdm>=4.65.0

# Optional: For enhanced logging and monitoring
colorlog>=6.7.0

